% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_classify.R, R/generics.R
\name{adaboost_classify}
\alias{adaboost_classify}
\alias{predict.mlpack_adaboost}
\title{AdaBoost Prediction}
\usage{
adaboost_classify(
  input_model,
  test,
  verbose = getOption("mlpack.verbose", FALSE)
)

\method{predict}{mlpack_adaboost}(object, newdata, type = c("predictions", "probabilities"), ...)
}
\arguments{
\item{input_model}{Input AdaBoost model (AdaBoostModel).}

\item{test}{Test dataset (numeric matrix).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}

\item{object}{An instantiated model object for which prediction is desired}

\item{newdata}{A test data set}

\item{type}{A character value selection predictions or probabilities}

\item{...}{Additional optional arguments affecting the prediction}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{Predicted labels for the test set (integer row).}
}
\description{
Class predictions from model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
