% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_probabilities.R
\name{adaboost_probabilities}
\alias{adaboost_probabilities}
\title{AdaBoost Probability Prediction}
\usage{
adaboost_probabilities(
  input_model,
  test,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Input AdaBoost model (AdaBoostModel).}

\item{test}{Test dataset (numeric matrix).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{probabilities}{Predicted class probabilities for each point in the
  test set (numeric matrix).}
}
\description{
Class probabilities from model.
}
\examples{
# \dontrun{ prob <- predict(model, newdata=X_test, type="probabilities") }
}
\author{
mlpack developers
}
