% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_linear_regression_predict.R
\name{bayesian_linear_regression_predict}
\alias{bayesian_linear_regression_predict}
\title{BayesianLinearRegression Prediction}
\usage{
bayesian_linear_regression_predict(
  input_model,
  test,
  stddevs = FALSE,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Trained BayesianLinearRegression model to use
(BayesianLinearRegression).}

\item{test}{Matrix containing points to regress on (test points)
(numeric matrix).}

\item{stddevs}{Return standard deviations along with predictions. 
Default value "FALSE" (logical).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{Matrix of predicted responses, with associated
  standard deviations if option selected (numeric matrix).}
}
\description{
An implementation of the Bayesian linear regression prediction: Given a
pre-trained model and a test data set, it provides model predictions.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
