% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_classify.R
\name{decision_tree_classify}
\alias{decision_tree_classify}
\title{Decision tree Prediction}
\usage{
decision_tree_classify(
  input_model,
  test,
  test_labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Pre-trained decision tree, to be used with test
points (DecisionTreeModel).}

\item{test}{Testing dataset (may contain categorical variables) (numeric
matrix/data.frame with info).}

\item{test_labels}{Test point labels, if accuracy calculation is desired
(integer row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{Class predictions for each test point (integer
  row).}
}
\description{
Class predictions from train decision tree model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
