% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoeffding_tree_classify.R
\name{hoeffding_tree_classify}
\alias{hoeffding_tree_classify}
\title{Hoeffding trees Classification}
\usage{
hoeffding_tree_classify(
  input_model,
  test,
  test_labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Input trained Hoeffding tree model
(HoeffdingTreeModel).}

\item{test}{Testing dataset (may be categorical) (numeric
matrix/data.frame with info).}

\item{test_labels}{Labels of test data (integer row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{Matrix to output label predictions for test data into
  (integer row).}
}
\description{
Class predictions from Hoeffding trees model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
