% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoeffding_tree_train.R
\name{hoeffding_tree_train}
\alias{hoeffding_tree_train}
\title{Hoeffding trees training}
\usage{
hoeffding_tree_train(
  training,
  batch_mode = FALSE,
  bins = 10,
  confidence = 0.95,
  info_gain = FALSE,
  labels = NA,
  max_samples = 5000,
  min_samples = 100,
  numeric_split_strategy = "binary",
  observations_before_binning = 100,
  passes = 1,
  test = NA,
  test_labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{training}{Training dataset (may be categorical) (numeric
matrix/data.frame with info).}

\item{batch_mode}{If true, samples will be considered in batch instead
of as a stream.  This generally results in better trees but at the cost of
memory usage and runtime.  Default value "FALSE" (logical).}

\item{bins}{If the 'domingos' split strategy is used, this specifies the
number of bins for each numeric split.  Default value "10" (integer).}

\item{confidence}{Confidence before splitting (between 0 and 1). 
Default value "0.95" (numeric).}

\item{info_gain}{If set, information gain is used instead of Gini
impurity for calculating Hoeffding bounds.  Default value "FALSE"
(logical).}

\item{labels}{Labels for training dataset (integer row).}

\item{max_samples}{Maximum number of samples before splitting.  Default
value "5000" (integer).}

\item{min_samples}{Minimum number of samples before splitting.  Default
value "100" (integer).}

\item{numeric_split_strategy}{The splitting strategy to use for numeric
features: 'domingos' or 'binary'.  Default value "binary" (character).}

\item{observations_before_binning}{If the 'domingos' split strategy is
used, this specifies the number of samples observed before binning is
performed.  Default value "100" (integer).}

\item{passes}{Number of passes to take over the dataset.  Default value
"1" (integer).}

\item{test}{Testing dataset (may be categorical) (numeric
matrix/data.frame with info).}

\item{test_labels}{Labels of test data (integer row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{output_model}{Output for trained Hoeffding tree model
  (HoeffdingTreeModel).}
}
\description{
An implementation of Hoeffding trees, a form of streaming decision tree for
classification.  Given labeled data a Hoeffding tree can be trained for later
use of predicting the classifications of new points.
}
\details{
Implements Hoeffding trees, a form of streaming decision tree suited best for
large (or streaming) datasets, supporting both categorical and numeric data. 
Given an input dataset, it is able to train the tree with numerous training
options, and return the model.

The training file and associated labels are specified with the "training" and
"labels" parameters, respectively. Optionally, if "labels" is not specified,
the labels are assumed to be the last dimension of the training dataset.

The training may be performed in batch mode (like a typical decision tree
algorithm) by specifying the "batch_mode" option, but this may not be the
best option for large datasets.
}
\examples{
# 
# 
# suppressMessages(library(mlpack)) # in case 'mlpack' is not yet loaded
# X <- as.matrix(read.csv("http://datasets.mlpack.org/iris.csv",
# header=FALSE))
# y <- as.matrix(read.csv("http://datasets.mlpack.org/iris_labels.csv",
# header=FALSE))
# pp <- preprocess_split(input=X, input_label=y, test_ratio=0.2)
# X_train <- pp[["training"]]
# X_test <- pp[["test"]]
# y_train <- pp[["training_labels"]]
# y_test <- pp[["test_labels"]]
# 
# model <- hoeffding_tree_train(training=X_train, labels=y_train)
# 
}
\author{
mlpack developers
}
