% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_svm_classify.R
\name{linear_svm_classify}
\alias{linear_svm_classify}
\title{Linear SVM Prediction}
\usage{
linear_svm_classify(
  input_model,
  test,
  test_labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Existing model (parameters) (LinearSVMModel).}

\item{test}{Matrix containing test dataset (numeric matrix).}

\item{test_labels}{Matrix containing test labels (integer row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{If test data is specified, this matrix is where the
  predictions for the test set will be saved (integer row).}
}
\description{
Class prediction from Linear SVM model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
