% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression_probabilities.R
\name{logistic_regression_probabilities}
\alias{logistic_regression_probabilities}
\title{L2-regularized Logistic Regression Probabilities}
\usage{
logistic_regression_probabilities(
  input_model,
  test,
  decision_boundary = 0.5,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Existing model (parameters) (LogisticRegression).}

\item{test}{Matrix containing test dataset (numeric matrix).}

\item{decision_boundary}{Decision boundary for prediction; if the
logistic function for a point is less than the boundary, the class is taken
to be 0; otherwise, the class is 1.  Default value "0.5" (numeric).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{probabilities}{Predicted class probabilities for each point in the
  test set (numeric matrix).}
}
\description{
An implementation of L2-regularized logistic regression for two-class
classification.  Uses a trained model to classify new points and provide
classification probabilities.
}
\examples{
# \dontrun{ prob <- predict(model, newdata=X_test, type="probabilities") }
}
\author{
mlpack developers
}
