% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mlpack}
\alias{mlpack}
\alias{mlpack-package}
\title{mlpack}
\description{
mlpack is a fast, flexible machine learning library, written in C++, that
aims to provide fast, extensible implementations of cutting-edge machine
learning algorithms. mlpack provides these algorithms as simple command-line
programs, C++ classes and bindings for : Python, Julia, Go and R which can
then be integrated into larger-scale machine learning solutions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.mlpack.org/doc/user/bindings/r.html}
  \item \url{https://github.com/mlpack/mlpack}
  \item Report bugs at \url{https://github.com/mlpack/mlpack/issues}
}

}
\author{
\strong{Maintainer}: Ryan Curtin \email{ryan@ratml.org} [contributor, copyright holder]

Authors:
\itemize{
  \item Ryan Curtin \email{ryan@ratml.org} [contributor, copyright holder]
  \item Yashwant Singh Parihar \email{yashwantsingh.sngh@gmail.com} [contributor, copyright holder]
  \item Dirk Eddelbuettel \email{edd@debian.org} [contributor, copyright holder]
  \item James Balamuta \email{james.balamuta@gmail.com} [contributor, copyright holder]
}

Other contributors:
\itemize{
  \item Bill March \email{march@gatech.edu} [contributor, copyright holder]
  \item Dongryeol Lee \email{dongryel@cc.gatech.edu} [contributor, copyright holder]
  \item Nishant Mehta \email{niche@cc.gatech.edu} [contributor, copyright holder]
  \item Parikshit Ram \email{p.ram@gatech.edu} [contributor, copyright holder]
  \item James Cline \email{james.cline@gatech.edu} [contributor, copyright holder]
  \item Sterling Peet \email{sterling.peet@gatech.edu} [contributor, copyright holder]
  \item Matthew Amidon \email{mamidon@gatech.edu} [contributor, copyright holder]
  \item Neil Slagle \email{npslagle@gmail.com} [contributor, copyright holder]
  \item Ajinkya Kale \email{kaleajinkya@gmail.com} [contributor, copyright holder]
  \item Vlad Grantcharov \email{vlad321@gatech.edu} [contributor, copyright holder]
  \item Noah Kauffman \email{notoriousnoah@gmail.com} [contributor, copyright holder]
  \item Rajendran Mohan \email{rmohan88@gatech.edu} [contributor, copyright holder]
  \item Trironk Kiatkungwanglai \email{trironk@gmail.com} [contributor, copyright holder]
  \item Patrick Mason \email{patrick.s.mason@gmail.com} [contributor, copyright holder]
  \item Marcus Edel \email{marcus.edel@fu-berlin.de} [contributor, copyright holder]
  \item Mudit Raj Gupta \email{mudit.raaj.gupta@gmail.com} [contributor, copyright holder]
  \item Sumedh Ghaisas \email{sumedhghaisas@gmail.com} [contributor, copyright holder]
  \item Michael Fox \email{michaelfox99@gmail.com} [contributor, copyright holder]
  \item Siddharth Agrawal \email{siddharth.950@gmail.com} [contributor, copyright holder]
  \item Saheb Motiani \email{saheb210692@gmail.com} [contributor, copyright holder]
  \item Yash Vadalia \email{yashdv@gmail.com} [contributor, copyright holder]
  \item Abhishek Laddha \email{laddhaabhishek11@gmail.com} [contributor, copyright holder]
  \item Vahab Akbarzadeh \email{v.akbarzadeh@gmail.com} [contributor, copyright holder]
  \item Andrew Wells \email{andrewmw94@gmail.com} [contributor, copyright holder]
  \item Zhihao Lou \email{lzh1984@gmail.com} [contributor, copyright holder]
  \item Udit Saxena \email{saxenda.udit@gmail.com} [contributor, copyright holder]
  \item Stephen Tu \email{tu.stephenl@gmail.com} [contributor, copyright holder]
  \item Jaskaran Singh \email{jaskaranvirdi@ymail.com} [contributor, copyright holder]
  \item Hritik Jain \email{hritik.jain.cse13@itbhu.ac.in} [contributor, copyright holder]
  \item Vladimir Glazachev \email{glazachev.vladimir@gmail.com} [contributor, copyright holder]
  \item QiaoAn Chen \email{kazenoyumechen@gmail.com} [contributor, copyright holder]
  \item Janzen Brewer \email{jahabrewer@gmail.com} [contributor, copyright holder]
  \item Trung Dinh \email{dinhanhtrung@gmail.com} [contributor, copyright holder]
  \item Tham Ngap Wei \email{thamngapwei@gmail.com} [contributor, copyright holder]
  \item Grzegorz Krajewski \email{krajekg@gmail.com} [contributor, copyright holder]
  \item Joseph Mariadassou \email{joe.mariadassou@gmail.com} [contributor, copyright holder]
  \item Pavel Zhigulin \email{pashaworking@gmail.com} [contributor, copyright holder]
  \item Andy Fang \email{AndyFang.DZ@gmail.com} [contributor, copyright holder]
  \item Barak Pearlmutter \email{barak+git@pearlmutter.net} [contributor, copyright holder]
  \item Ivari Horm \email{ivari@risk.ee} [contributor, copyright holder]
  \item Dhawal Arora \email{d.p.arora1@gmail.com} [contributor, copyright holder]
  \item Alexander Leinoff \email{alexander-leinoff@uiowa.edu} [contributor, copyright holder]
  \item Palash Ahuja \email{abhor902@gmail.com} [contributor, copyright holder]
  \item Yannis Mentekidis \email{mentekid@gmail.com} [contributor, copyright holder]
  \item Ranjan Mondal \email{ranjan.rev@gmail.com} [contributor, copyright holder]
  \item Mikhail Lozhnikov \email{lozhnikovma@gmail.com} [contributor, copyright holder]
  \item Marcos Pividori \email{marcos.pividori@gmail.com} [contributor, copyright holder]
  \item Keon Kim \email{kwk236@gmail.com} [contributor, copyright holder]
  \item Nilay Jain \email{nilayjain13@gmail.com} [contributor, copyright holder]
  \item Peter Lehner \email{peter.lehner@dlr.de} [contributor, copyright holder]
  \item Anuraj Kanodia \email{akanuraj200@gmail.com} [contributor, copyright holder]
  \item Ivan Georgiev \email{ivan@jonan.info} [contributor, copyright holder]
  \item Shikhar Bhardwaj \email{shikharbhardwaj68@gmail.com} [contributor, copyright holder]
  \item Yashu Seth \email{yashuseth2503@gmail.com} [contributor, copyright holder]
  \item Mike Izbicki \email{mike@izbicki.me} [contributor, copyright holder]
  \item Sudhanshu Ranjan \email{sranjan.sud@gmail.com} [contributor, copyright holder]
  \item Piyush Jaiswal \email{piyush.jaiswal@st.niituniversity.in} [contributor, copyright holder]
  \item Dinesh Raj \email{dinu.iota@gmail.com} [contributor, copyright holder]
  \item Vivek Pal \email{vivekpal.dtu@gmail.com} [contributor, copyright holder]
  \item Prasanna Patil \email{prasannapatil08@gmail.com} [contributor, copyright holder]
  \item Lakshya Agrawal \email{zeeshan.lakshya@gmail.com} [contributor, copyright holder]
  \item Praveen Ch \email{chvsp972911@gmail.com} [contributor, copyright holder]
  \item Kirill Mishchenko \email{ki.mishchenko@gmail.com} [contributor, copyright holder]
  \item Abhinav Moudgil \email{abhinavmoudgil95@gmail.com} [contributor, copyright holder]
  \item Thyrix Yang \email{thyrixyang@gmail.com} [contributor, copyright holder]
  \item Sagar B Hathwar \email{sagarbhathwar@gmail.com} [contributor, copyright holder]
  \item Nishanth Hegde \email{hegde.nishanth@gmail.com} [contributor, copyright holder]
  \item Parminder Singh \email{parmsingh101@gmail.com} [contributor, copyright holder]
  \item CodeAi \email{benjamin.bales@assrc.us} [contributor, copyright holder]
  \item Franciszek Stokowacki \email{franek.stokowacki@gmail.com} [contributor, copyright holder]
  \item Samikshya Chand \email{samikshya289@gmail.com} [contributor, copyright holder]
  \item N Rajiv Vaidyanathan \email{rajivvaidyanathan4@gmail.com} [contributor, copyright holder]
  \item Kartik Nighania \email{kartiknighania@gmail.com} [contributor, copyright holder]
  \item Eugene Freyman \email{evg.freyman@gmail.com} [contributor, copyright holder]
  \item Manish Kumar \email{manish887kr@gmail.com} [contributor, copyright holder]
  \item Haritha Sreedharan Nair \email{haritha1313@gmail.com} [contributor, copyright holder]
  \item Sourabh Varshney \email{sourabhvarshney111@gmail.com} [contributor, copyright holder]
  \item Projyal Dev \email{projyal@gmail.com} [contributor, copyright holder]
  \item Nikhil Goel \email{nikhilgoel199797@gmail.com} [contributor, copyright holder]
  \item Shikhar Jaiswal \email{jaiswalshikhar87@gmail.com} [contributor, copyright holder]
  \item B Kartheek Reddy \email{bkartheekreddy@gmail.com} [contributor, copyright holder]
  \item Atharva Khandait \email{akhandait45@gmail.com} [contributor, copyright holder]
  \item Wenhao Huang \email{wenhao.huang.work@gmail.com} [contributor, copyright holder]
  \item Roberto Hueso \email{robertohueso96@gmail.com} [contributor, copyright holder]
  \item Prabhat Sharma \email{prabhatsharma7298@gmail.com} [contributor, copyright holder]
  \item Tan Jun An \email{yamidarkxxx@gmail.com} [contributor, copyright holder]
  \item Moksh Jain \email{mokshjn00@gmail.com} [contributor, copyright holder]
  \item Manthan-R-Sheth \email{manthanrsheth96@gmail.com} [contributor, copyright holder]
  \item Namrata Mukhija \email{namratamukhija@gmail.com} [contributor, copyright holder]
  \item Conrad Sanderson [contributor, copyright holder]
  \item Thanasis Mattas \email{mattasa@auth.gr} [contributor, copyright holder]
  \item Shashank Shekhar \email{contactshashankshekhar@gmail.com} [contributor, copyright holder]
  \item Yasmine Dumouchel \email{yasmine.dumouchel@gmail.com} [contributor, copyright holder]
  \item German Lancioni [contributor, copyright holder]
  \item Arash Abghari \email{arash.abghari@gmail.com} [contributor, copyright holder]
  \item Ayush Chamoli [contributor, copyright holder]
  \item Tommi Laivamaa \email{tommi.laivamaa@protonmail.com} [contributor, copyright holder]
  \item Kim SangYeon \email{sy0814k@gmail.com} [contributor, copyright holder]
  \item Niteya Shah \email{niteya.56@gmail.com} [contributor, copyright holder]
  \item Toshal Agrawal \email{tagrawal1339@gmail.com} [contributor, copyright holder]
  \item Dan Timson [contributor, copyright holder]
  \item Miguel Canteras \email{mcanteras@gmail.com} [contributor, copyright holder]
  \item Bishwa Karki \email{karkeebishwa1@gmail.com} [contributor, copyright holder]
  \item Mehul Kumar Nirala \email{mehulkumarnirala@gmail.com} [contributor, copyright holder]
  \item Heet Sankesara \email{heetsankesara3@gmail.com} [contributor, copyright holder]
  \item Jeffin Sam \email{sam.jeffin@gmail.com} [contributor, copyright holder]
  \item Vikas S Shetty \email{shettyvikas209@gmail.com} [contributor, copyright holder]
  \item Khizir Siddiqui \email{khizirsiddiqui@gmail.com} [contributor, copyright holder]
  \item Tejasvi Tomar \email{tstomar@outlook.com} [contributor, copyright holder]
  \item Jai Agarwal \email{jai.bhageria@gmail.com} [contributor, copyright holder]
  \item Ziyang Jiang \email{zij004@alumni.stanford.edu} [contributor, copyright holder]
  \item Rohit Kartik \email{rohit.audrey@gmail.com} [contributor, copyright holder]
  \item Aditya Viki \email{adityaviki01@gmail.com} [contributor, copyright holder]
  \item Kartik Dutt \email{kartikdutt@live.in} [contributor, copyright holder]
  \item Suryoday Basak \email{suryodaybasak@gmail.com} [contributor, copyright holder]
  \item Sriram S K \email{sriramsk1999@gmail.com} [contributor, copyright holder]
  \item Manoranjan Kumar Bharti ( Nakul Bharti ) \email{knakul853@gmail.com} [contributor, copyright holder]
  \item Saraansh Tandon \email{saraanshtandon1999@gmail.com} [contributor, copyright holder]
  \item Gaurav Singh \email{gs8763076@gmail.com} [contributor, copyright holder]
  \item Lakshya Ojha \email{ojhalakshya@gmail.com} [contributor, copyright holder]
  \item Bisakh Mondal \email{bisakhmondal00@gmail.com} [contributor, copyright holder]
  \item Benson Muite \email{benson_muite@emailplus.org} [contributor, copyright holder]
  \item Sarthak Bhardwaj \email{7sarthakbhardwaj@gmail.com} [contributor, copyright holder]
  \item Aakash Kaushik \email{kaushikaakash7539@gmail.com} [contributor, copyright holder]
  \item Anush Kini \email{anushkini@gmail.com} [contributor, copyright holder]
  \item Nippun Sharma \email{inbox.nippun@gmail.com} [contributor, copyright holder]
  \item Rishabh Garg \email{rishabhgarg108@gmail.com} [contributor, copyright holder]
  \item Sudhakar Brar \email{dxhrmhall1449@tutanota.com} [contributor, copyright holder]
  \item Alex Nguyen \email{alexvn.edu@gmail.com} [contributor, copyright holder]
  \item Gaurav Ghati \email{gauravghatii@gmail.com} [contributor, copyright holder]
  \item Anmolpreet Singh \email{anmol323c@gmail.com} [contributor, copyright holder]
  \item Anjishnu Mukherjee \email{amukher6@gmu.edu} [contributor, copyright holder]
  \item Omar Shrit \email{omar@shrit.me} [contributor, copyright holder]
  \item Tru Hoang \email{trugiahoang@gmail.com} [contributor, copyright holder]
  \item Mark Fischinger \email{markfischinger@gmail.com} [contributor, copyright holder]
  \item Muhammad Fawwaz Mayda \email{maydafawwaz@gmail.com} [contributor, copyright holder]
  \item Roshan Nrusing Swain \email{swainroshan001@gmail.com} [contributor, copyright holder]
  \item Suvarsha Chennareddy \email{suvarshachennareddy@gmail.com} [contributor, copyright holder]
  \item Shubham Agrawal \email{shubham.agra1206@gmail.com} [contributor, copyright holder]
  \item Sri Madhan M \email{srimadhan11@gmail.com} [contributor, copyright holder]
  \item Zhuojin Liu \email{zhuojinliu.cs@gmail.com} [contributor, copyright holder]
  \item Richèl Bilderbeek \email{richel@richelbilderbeek.nl} [contributor, copyright holder]
  \item Chetan Pandey \email{chetanpandey1266@gmail.com} [contributor, copyright holder]
  \item Nikolay Apanasov \email{nikolay@apanasov.org} [contributor, copyright holder]
  \item Martin Lambertsen \email{github@lambertsen.one} [contributor, copyright holder]
  \item Andrea Novellini \email{andre.novellini@gmail.com} [contributor, copyright holder]
  \item Felix Patschkowski \email{felix.patschkowski@gmail.com} [contributor, copyright holder]
  \item Benjamin A. Beasley \email{code@musicinmybrain.net} [contributor, copyright holder]
  \item Maksym Prots \email{imaxprots@gmail.com} [contributor, copyright holder]
  \item Zachary Ng \email{zachn716@gmail.com} [contributor, copyright holder]
  \item Ranjodh Singh \email{ranjodhsingh1729@gmail.com} [contributor, copyright holder]
  \item Mohammad Mundiwala \email{mohammadmundiwala@gmail.com} [contributor, copyright holder]
}

}
