% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbc_probabilities.R
\name{nbc_probabilities}
\alias{nbc_probabilities}
\title{Parametric Naive Bayes Classifier Probabilities}
\usage{
nbc_probabilities(
  input_model,
  test,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Input Naive Bayes model (NBCModel).}

\item{test}{A matrix containing the test set (numeric matrix).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{probabilities}{The matrix in which the predicted probability of
  labels for the test set will be written (numeric matrix).}
}
\description{
Class probabilities from a Naive Bayes Classifier model.
}
\examples{
# \dontrun{ prob <- predict(model, newdata=X_test, type="probabilities") }
}
\author{
mlpack developers
}
