% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbc_train.R
\name{nbc_train}
\alias{nbc_train}
\title{Parametric Naive Bayes Classifier training}
\usage{
nbc_train(
  training,
  incremental_variance = FALSE,
  labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{training}{A matrix containing the training set (numeric matrix).}

\item{incremental_variance}{The variance of each class will be
calculated incrementally.  Default value "FALSE" (logical).}

\item{labels}{A vector containing labels for the training set (integer
row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{output_model}{File to save trained Naive Bayes model to
  (NBCModel).}
}
\description{
An implementation of the Naive Bayes Classifier, used for classification.
Given labeled data, an NBC model is be trained for later use for
classification on new data.
}
\details{
Implements the Naive Bayes classifier on the given labeled training set for
us of that trained model to classify the points in a given test set.

The training set is specified with the "training" parameter.  Labels may be
either the last row of the training set, or alternately the "labels"
parameter may be specified to pass a separate matrix of labels.

The "incremental_variance" parameter can be used to force the training to use
an incremental algorithm for calculating variance.  This is slower, but can
help avoid loss of precision in some cases.
}
\examples{
# 
# 
# suppressMessages(library(mlpack)) # in case 'mlpack' is not yet loaded
# X <- as.matrix(read.csv("http://datasets.mlpack.org/iris.csv",
# header=FALSE))
# y <- as.matrix(read.csv("http://datasets.mlpack.org/iris_labels.csv",
# header=FALSE))
# pp <- preprocess_split(input=X, input_label=y, test_ratio=0.2)
# X_train <- pp[["training"]]
# X_test <- pp[["test"]]
# y_train <- pp[["training_labels"]]
# y_test <- pp[["test_labels"]]
# 
# model <- nbc_train(training=X_train, labels=y_train)
# 
}
\author{
mlpack developers
}
