% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron_classify.R
\name{perceptron_classify}
\alias{perceptron_classify}
\title{Perceptron Prediction}
\usage{
perceptron_classify(
  input_model,
  test,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Input perceptron model (PerceptronModel).}

\item{test}{A matrix containing the test set (numeric matrix).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{The matrix in which the predicted labels for the test
  set will be written (integer row).}
}
\description{
Class predictions from perceptron model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
