% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron_train.R
\name{perceptron_train}
\alias{perceptron_train}
\title{Perceptron training}
\usage{
perceptron_train(
  training,
  labels = NA,
  max_iterations = 1000,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{training}{A matrix containing the training set (numeric matrix).}

\item{labels}{A matrix containing labels for the training set (integer
row).}

\item{max_iterations}{The maximum number of iterations the perceptron is
to be ru.  Default value "1000" (integer).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{output_model}{Output for trained perceptron model
  (PerceptronModel).}
}
\description{
An implementation of a perceptron---a single level neural network---for
classification.  Given labeled data, a perceptron can be trained and later be
used for classification on new points.
}
\details{
Implementation of a perceptron, which is a single level neural network. The
perceptron makes its predictions based on a linear predictor function
combining a set of weights with the feature vector.  The perceptron learning
rule is able to converge, given enough iterations (specified using the
"max_iterations" parameter), if the data supplied is linearly separable.  The
perceptron is parameterized by a matrix of weight vectors that denote the
numerical weights of the neural network.
}
\examples{
# 
# 
# suppressMessages(library(mlpack)) # in case 'mlpack' is not yet loaded
# X <- as.matrix(read.csv("http://datasets.mlpack.org/iris.csv",
# header=FALSE))
# y <- as.matrix(read.csv("http://datasets.mlpack.org/iris_labels.csv",
# header=FALSE))
# pp <- preprocess_split(input=X, input_label=y, test_ratio=0.2)
# X_train <- pp[["training"]]
# X_test <- pp[["test"]]
# y_train <- pp[["training_labels"]]
# y_test <- pp[["test_labels"]]
# 
# model <- perceptron_train(training=X_train, labels=y_train,
#   max_iterations=100)
#   
}
\author{
mlpack developers
}
