% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_classify.R
\name{random_forest_classify}
\alias{random_forest_classify}
\title{Random Forests Prediction}
\usage{
random_forest_classify(
  input_model,
  test,
  test_labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Pre-trained random forest to use for classification
(RandomForestModel).}

\item{test}{Test dataset to produce predictions for (numeric matrix).}

\item{test_labels}{Test dataset labels, if accuracy calculation is
desired (integer row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{Predicted classes for each point in the test set
  (integer row).}
}
\description{
Class predictions from random forest model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
