% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_probabilities.R
\name{random_forest_probabilities}
\alias{random_forest_probabilities}
\title{Random Forests Probabilities}
\usage{
random_forest_probabilities(
  input_model,
  test,
  test_labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Pre-trained random forest to use for classification
(RandomForestModel).}

\item{test}{Test dataset to produce predictions for (numeric matrix).}

\item{test_labels}{Test dataset labels, if accuracy calculation is
desired (integer row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{probabilities}{Predicted class probabilities for each point in the
  test set (numeric matrix).}
}
\description{
Class probabilities from random forest model.
}
\examples{
# \dontrun{ prob <- predict(model, newdata=X_test, type="probabilities") }
}
\author{
mlpack developers
}
