% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softmax_regression_probabilities.R
\name{softmax_regression_probabilities}
\alias{softmax_regression_probabilities}
\title{Softmax Regression Classification}
\usage{
softmax_regression_probabilities(
  input_model,
  test,
  test_labels = NA,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Existing model (parameters) (SoftmaxRegression).}

\item{test}{Matrix containing test dataset (numeric matrix).}

\item{test_labels}{Matrix containing test labels (integer row).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{probabilities}{Matrix to save class probabilities for test dataset
  into (numeric matrix).}
}
\description{
An implementation of softmax regression for classification, which is a
multiclass generalization of logistic regression.  Given a pre-trained
softmax regression model, new points are classified.
}
\examples{
# \dontrun{ prob <- predict(model, newdata=X_test, type="probabilities") }
}
\author{
mlpack developers
}
