% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost_classify.R
\name{adaboost_classify}
\alias{adaboost_classify}
\title{AdaBoost Prediction}
\usage{
adaboost_classify(
  input_model,
  test,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Input AdaBoost model (AdaBoostModel).}

\item{test}{Test dataset (numeric matrix).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{Predicted labels for the test set (integer row).}
}
\description{
Class predictions from model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
