% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lars_predict.R
\name{lars_predict}
\alias{lars_predict}
\title{LARS Prediction}
\usage{
lars_predict(input_model, test, verbose = getOption("mlpack.verbose", FALSE))
}
\arguments{
\item{input_model}{Trained LARS model to use (LARS).}

\item{test}{Matrix containing points to regress on (test points)
(numeric matrix).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{predictions}{Matrix containing predicted responses (numeric
  matrix).}
}
\description{
An implementation of Least Angle Regression (stagewise/lasso), also known as
LARS.  This program can use a pre-trained LARS/LASSO/Elastic Net model to
output regression predictions from a test set.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
