% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_predict.R
\name{linear_regression_predict}
\alias{linear_regression_predict}
\title{Linear Regression Prediction}
\usage{
linear_regression_predict(
  input_model,
  test,
  verbose = getOption("mlpack.verbose", FALSE)
)
}
\arguments{
\item{input_model}{Existing LinearRegression model to use
(LinearRegression).}

\item{test}{Matrix containing X' (test regressors) (numeric matrix).}

\item{verbose}{Display informational messages and the full list of
parameters and timers at the end of execution.  Default value
"getOption("mlpack.verbose", FALSE)" (logical).}
}
\value{
A list with several components defining the class attributes:
\item{output_predictions}{Matrix containing predicted responses (numeric
  row).}
}
\description{
Predictions from model.
}
\examples{
# \dontrun{ pred <- predict(model, newdata=X_test) }
}
\author{
mlpack developers
}
